refactor(hypervisor, scheduler): moved helper and data structs to submodules
